extract Tabs
